Quiesce all remaining g++/clang++ warnings (-Wall) in libbiomeval on OS X.
[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 2941]

+++ src/libbiomeval/be_image_jpeg.cpp
@@ -163,7 +163,7 @@ BiometricEvaluation::Image::JPEG::getRawGrayscaleData(
 			 *
 			 * TODO: Use a colormap to support 2-7 bit depth.
 			 */
-			for (int i = 0; i < row_stride; i++)
+			for (uint64_t i = 0; i < row_stride; i++)
 				if (buffer[0][i] == 0x01)
 					buffer[0][i] = 0xFF;
 			break;
